requirements.txt
PySide6>=6.4.0
mitmproxy==10.1.5
dnspython>=2.3.0
requests>=2.28.0
beautifulsoup4>=4.11.0
python-nmap>=0.7.1
SQLAlchemy>=2.0.0
jinja2>=3.1.0
pdfkit>=1.0.0
cryptography>=41.0.0
aiohttp>=3.8.0
asyncio>=3.4.3
pyjwt>=2.8.0
pyyaml==6.0.1
tqdm==4.66.1
colorama==0.4.6
passlib==1.7.4
pyinstaller>=5.13.0
pyupdater>=4.0 